// src/app/admin-view/models/subExpense.ts
export class SubExpenses {
    Description: string;
    Distid: string;
    EID: string;
    ElectionType: string;
    InsertedBy: string;
    InsertedDate: string;
    IsActive: string;
    LocalBodyId: string;
    MeasuringUnit: string;
    NotFound: string;
    SEID: string;
    StandardRate: string;
    StateId: string;
    Status: string;
    SubExpenseType: string;
}